==> 1B-UPC_parse.R <==
# R script: 1B-UPC_parse.R
# Author: Inge Seim
# Description: This script will parse SCAN.UPC data (generated by SCAN.UPC.R) and output a data frame containing the UPC scores of probe sets of interest.


# remove all objects
rm(list = ls(all = TRUE))

# go to a directory with your UPC data (ExpressionSet object converted to a data frame previously (df.x <- as.data.frame(exprs(ExpressionSet.x)). See Supplementary Note 2)
setwd("/Volumes/ExonArray/GHSROS/data/UPCDF/")

# list directories
startingdirectory <- getwd()
UPCdirs <- list.dirs('.', recursive=FALSE,full.names=FALSE)
length(UPCdirs)

# create empty data frame to feed data from the loop into
create_empty_table <- function(num_rows, num_cols) {
frame <- data.frame(matrix(NA, nrow = num_rows, ncol = num_cols))
return(frame)
}
mergedDF <- create_empty_table(1,0)

# Decide on probe set(s) of interest here so that one does not have to load all 1,411,399 probe sets into memory for thousands of samples
# 2652604 = GHSROS  
# 2705695 = GHSR extended exon 1 = exon 1b (GHSR1b-only)
# 2705696 = GHSR exon 1 (common to both GHSR1a and GHSR1b)
# 2705694 = GHSR exon 2 (common to both GHSR1a and GHSR1b)
targetprobesets <- c("2652604","2705694","2705695","2705696")

# open each UPC data file and merge
for(i in 1:length(UPCdirs))
{   # START OF OUTER	 LOOP

print(UPCdirs[i])
filedir <- paste("./",UPCdirs[i],sep="")
files <- list.files(filedir, recursive=FALSE,full.names=FALSE,pattern = ".*UPC*DF.*\\.rds")
for(i in 1:length(files))
{             # START OF INNER LOOP
print(files[i])
opentemp <- paste(filedir,"/",files[i],sep="")
temp <- readRDS(file=opentemp)
temp <- temp[c(targetprobesets),] #
mergedDF <- cbind(mergedDF, temp)
} # END OF INNER LOOP
} # END OF OUTER LOOP

# save output
saveRDS(mergedDF,file="../UPC.mergedDF.rds")
# ~~~~~~~~~~~~~~~~~~~~~~~~ THIS IS THE END ~~~~~~~~~~~~~~~~~~~~~~~~ #
